Update user data provider
