+++ src/types/api.d.ts
@@ -15,6 +15,7 @@ type userDocument = {
   why_pm?: string;
   returning_member?: "yes" | "no";
   paymentVerified?: boolean;
+  isGuest?: boolean;
 };
 
 type loginBody = {
