+++ src/vite-env.d.ts
@@ -2,7 +2,9 @@
 
 interface ImportMetaEnv {
     readonly VITE_API_URL: string;
+    readonly VITE_AUTH0_CLIENT_ID: string;
     readonly VITE_AUTH0_API_URL: string;
+    readonly VITE_AUTH0_API_ACCESS_TOKEN: string;
 }
 
 interface ImportMeta {
